refactor(record): record index 추가
