+++ app/(user)/LikeDocsList.tsx
@@ -1,7 +1,7 @@
 import Accordion from "@/components/Accordion";
 import { ContributeDocsType } from "@/types/contributeDocs.interface";
 import Link from "next/link";
-import { CLASSIFY } from "@/record/docsType.record";
+import { CLASSIFY } from "@/record";
 import * as styles from "./style.css";
 
 const LikeDocsList = ({ likeList }: { likeList: Array<ContributeDocsType> }) => {
