+++ app/coin/TradeHistory.tsx
@@ -6,7 +6,7 @@ import Image from "next/image";
 import { useCancelTradeMutation } from "@/services/coin/coin.mutation";
 import { toast } from "react-toastify";
 import { useDate } from "@/hooks";
-import { TRADE } from "@/record/trade.record";
+import { TRADE } from "@/record";
 import Toastify from "@/components/Toastify";
 import * as styles from "./TradeHistory.css";
