+++ app/coin/trade/[accountId]/TradeHistory.tsx
@@ -7,7 +7,7 @@ import { priceComma } from "@/utils";
 import Image from "next/image";
 import Container from "@/components/Container";
 import { useDate } from "@/hooks";
-import { TRADE } from "@/record/trade.record";
+import { TRADE } from "@/record";
 import * as styles from "../../TradeHistory.css";
 
 const Trade: FC<{ accountId: number }> = ({ accountId }) => {
