+++ components/Container/index.tsx
@@ -7,7 +7,7 @@ import { useRouter } from "next/navigation";
 import { toast } from "react-toastify";
 import { useQueryClient } from "@tanstack/react-query";
 import { docsQuery } from "@/services/docs/docs.query";
-import { CLASSIFY } from "@/record/docsType.record";
+import { CLASSIFY } from "@/record";
 import { useDate, useModal, useUser } from "@/hooks";
 import * as styles from "./style.css";
 import Toastify from "../Toastify";
