+++ record/index.ts
@@ -1,0 +1,3 @@
+export { default as CLASSIFY } from "./classify.record";
+export { default as ROLE } from "./role.record";
+export { default as TRADE } from "./trade.record";
