+++ record/role.record.ts
@@ -1,5 +1,7 @@
-export const ROLE = {
+const ROLE = {
   ADMIN: "관리자",
   READONLY: "읽기전용 사용자",
   USER: "유저",
 } as Record<string, string>;
+
+export default ROLE;
